# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Will Be Done API",
    "description": "Public HTTP API for Will Be Done.",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Token returned by the login or register endpoint."
      }
    },
...
  },
  "paths": {
...
    "/api/v1/spaces/{spaceId}/tasks/{taskId}/checklist-items": {
      "get": {
        "operationId": "listTaskChecklistItems",
        "summary": "List a task's checklist items",
        "tags": ["Checklist items"],
        "parameters": [
          {
            "schema": { "type": "string", "minLength": 1 },
            "in": "path",
            "name": "spaceId",
            "required": true,
            "description": "Space identifier"
          },
          {
            "schema": { "type": "string", "minLength": 1 },
            "in": "path",
            "name": "taskId",
            "required": true,
            "description": "Task identifier"
          }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "checklistItems": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "type": {
                            "type": "string",
                            "enum": ["checklistItem"]
                          },
                          "id": {
                            "type": "string",
                            "description": "Checklist item identifier"
                          },
                          "parentId": {
                            "type": "string",
                            "description": "Task or task-template identifier"
                          },
                          "parentType": {
                            "type": "string",
                            "enum": ["task", "template"]
                          },
                          "state": {
                            "type": "string",
                            "enum": ["todo", "done"]
                          },
                          "content": { "type": "string" },
                          "createdAt": {
                            "type": "integer",
                            "minimum": 0,
                            "maximum": 9007199254740991
                          },
                          "checkedAt": {
                            "anyOf": [
                              {
                                "type": "integer",
                                "minimum": 0,
                                "maximum": 9007199254740991
                              },
                              { "type": "null" }
                            ]
                          }
                        },
                        "required": [
                          "type",
                          "id",
                          "parentId",
                          "parentType",
                          "state",
                          "content",
                          "createdAt",
                          "checkedAt"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": ["checklistItems"],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "BAD_REQUEST",
                        "UNAUTHORIZED",
                        "FORBIDDEN",
                        "NOT_FOUND",
                        "CONFLICT",
                        "INTERNAL_SERVER_ERROR"
                      ]
                    },
                    "message": { "type": "string" }
                  },
                  "required": ["code", "message"],
                  "additionalProperties": false,
                  "description": "Error response"
                }
              }
            }
          },
          "403": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "BAD_REQUEST",
                        "UNAUTHORIZED",
                        "FORBIDDEN",
                        "NOT_FOUND",
                        "CONFLICT",
                        "INTERNAL_SERVER_ERROR"
                      ]
                    },
                    "message": { "type": "string" }
                  },
                  "required": ["code", "message"],
                  "additionalProperties": false,
                  "description": "Error response"
                }
              }
            }
          },
          "404": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "BAD_REQUEST",
                        "UNAUTHORIZED",
                        "FORBIDDEN",
                        "NOT_FOUND",
                        "CONFLICT",
                        "INTERNAL_SERVER_ERROR"
                      ]
                    },
                    "message": { "type": "string" }
                  },
                  "required": ["code", "message"],
                  "additionalProperties": false,
                  "description": "Error response"
                }
              }
            }
          },
          "500": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "BAD_REQUEST",
                        "UNAUTHORIZED",
                        "FORBIDDEN",
                        "NOT_FOUND",
                        "CONFLICT",
                        "INTERNAL_SERVER_ERROR"
                      ]
                    },
                    "message": { "type": "string" }
                  },
                  "required": ["code", "message"],
                  "additionalProperties": false,
                  "description": "Error response"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createTaskChecklistItem",
        "summary": "Create a checklist item for a task",
        "tags": ["Checklist items"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": { "type": "string" },
                  "state": { "type": "string", "enum": ["todo", "done"] },
                  "placement": {
                    "oneOf": [
                      {
                        "type": "object",
                        "properties": {
                          "kind": { "type": "string", "enum": ["first"] }
                        },
                        "required": ["kind"],
                        "additionalProperties": false
                      },
                      {
                        "type": "object",
                        "properties": {
                          "kind": { "type": "string", "enum": ["last"] }
                        },
                        "required": ["kind"],
                        "additionalProperties": false
                      },
                      {
                        "type": "object",
                        "properties": {
                          "kind": { "type": "string", "enum": ["before"] },
                          "anchorId": {
                            "type": "string",
                            "minLength": 1,
                            "description": "Sibling to place the entity before"
                          }
                        },
                        "required": ["kind", "anchorId"],
                        "additionalProperties": false
                      },
                      {
                        "type": "object",
                        "properties": {
                          "kind": { "type": "string", "enum": ["after"] },
                          "anchorId": {
                            "type": "string",
                            "minLength": 1,
                            "description": "Sibling to place the entity after"
                          }
                        },
                        "required": ["kind", "anchorId"],
                        "additionalProperties": false
                      }
                    ],
                    "description": "Position within an ordered collection"
                  }
                },
                "required": ["content"],
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": { "type": "string", "minLength": 1 },
            "in": "path",
            "name": "spaceId",
            "required": true,
            "description": "Space identifier"
          },
          {
            "schema": { "type": "string", "minLength": 1 },
            "in": "path",
            "name": "taskId",
            "required": true,
            "description": "Task identifier"
          }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "checklistItem": {
                      "type": "object",
                      "properties": {
                        "type": { "type": "string", "enum": ["checklistItem"] },
                        "id": {
                          "type": "string",
                          "description": "Checklist item identifier"
                        },
                        "parentId": {
                          "type": "string",
                          "description": "Task or task-template identifier"
                        },
                        "parentType": {
                          "type": "string",
                          "enum": ["task", "template"]
                        },
                        "state": { "type": "string", "enum": ["todo", "done"] },
                        "content": { "type": "string" },
                        "createdAt": {
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 9007199254740991
                        },
                        "checkedAt": {
                          "anyOf": [
                            {
                              "type": "integer",
                              "minimum": 0,
                              "maximum": 9007199254740991
                            },
                            { "type": "null" }
                          ]
                        }
                      },
                      "required": [
                        "type",
                        "id",
                        "parentId",
                        "parentType",
                        "state",
                        "content",
                        "createdAt",
                        "checkedAt"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": ["checklistItem"],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "BAD_REQUEST",
                        "UNAUTHORIZED",
                        "FORBIDDEN",
                        "NOT_FOUND",
                        "CONFLICT",
                        "INTERNAL_SERVER_ERROR"
                      ]
                    },
                    "message": { "type": "string" }
                  },
                  "required": ["code", "message"],
                  "additionalProperties": false,
                  "description": "Error response"
                }
              }
            }
          },
          "403": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "BAD_REQUEST",
                        "UNAUTHORIZED",
                        "FORBIDDEN",
                        "NOT_FOUND",
                        "CONFLICT",
                        "INTERNAL_SERVER_ERROR"
                      ]
                    },
                    "message": { "type": "string" }
                  },
                  "required": ["code", "message"],
                  "additionalProperties": false,
                  "description": "Error response"
                }
              }
            }
          },
          "404": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "BAD_REQUEST",
                        "UNAUTHORIZED",
                        "FORBIDDEN",
                        "NOT_FOUND",
                        "CONFLICT",
                        "INTERNAL_SERVER_ERROR"
                      ]
                    },
                    "message": { "type": "string" }
                  },
                  "required": ["code", "message"],
                  "additionalProperties": false,
                  "description": "Error response"
                }
              }
            }
          },
          "409": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "BAD_REQUEST",
                        "UNAUTHORIZED",
                        "FORBIDDEN",
                        "NOT_FOUND",
                        "CONFLICT",
                        "INTERNAL_SERVER_ERROR"
                      ]
                    },
                    "message": { "type": "string" }
                  },
                  "required": ["code", "message"],
                  "additionalProperties": false,
                  "description": "Error response"
                }
              }
            }
          },
          "500": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "BAD_REQUEST",
                        "UNAUTHORIZED",
                        "FORBIDDEN",
                        "NOT_FOUND",
                        "CONFLICT",
                        "INTERNAL_SERVER_ERROR"
                      ]
                    },
                    "message": { "type": "string" }
                  },
                  "required": ["code", "message"],
                  "additionalProperties": false,
                  "description": "Error response"
                }
              }
            }
          }
        }
      }
    },
...
  }
}